reorganising React project structure
