--- frontend/src/components/buttons/LoginButton.tsx
@@ -1,9 +1,0 @@
-const LoginButton = () => {
-	return (
-		<button className="px-4 py-2 bg-blue-500 hover:bg-blue-600 text-white text-sm font-medium rounded-full">
-			Sign in
-		</button>
-	);
-};
-
-export default LoginButton;
