+++ frontend/src/pages/dashboard/Papers.tsx
@@ -1,0 +1,1 @@
+export {};
